requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
backports.tarfile==1.2.0
Brotli==1.1.0
brotlicffi==1.1.0.0
CacheControl==0.14.0
calculate-builder==3.4.0a1
calculate-client==3.2.2
calculate-console==3.2.2
calculate-console-gui==3.5.0a1
calculate-core==3.5.7
calculate-desktop==3.2.2
calculate-install==3.2.2
calculate-lib==3.2.2
calculate-update==3.7.3
certifi-system-store==3021.3.16
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
cupshelpers==1.0
distlib==0.3.8
distro==1.9.0
Django==5.0.7
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.21.2
frozenlist==1.4.1
gpg==1.23.2
idna==3.7
imagesize==1.4.1
installer==0.7.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.0
Jinja2==3.1.4
linkify-it-py==2.0.3
lxml==5.2.2
M2Crypto==0.41.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
multidict==6.0.5
notify2==0.3.1
olefile==0.47
ordered-set==4.1.0
packaging==24.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
ply==3.11
plyvel==1.5.1
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.26.0
pycparser==2.22
pycryptodome==3.20.0
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.46.0
pyinotify==0.9.6
pyOpenSSL==24.1.0
pyproject_hooks==1.1.0
PyQt5==5.15.10
PyQt5_sip==12.15.0
PySocks==1.7.1
python-ldap==3.4.4
pytz==2024.1
reportlab==4.2.0
requests==2.32.3
resolvelib==1.0.1
rich==13.7.1
setuptools-scm==8.1.0
six==1.16.0
sjcl==0.2.1
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx_celery==2.1.3
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.1
tenacity==8.5.0
tomli==2.0.1
trove-classifiers==2024.7.2
truststore==0.9.1
uc-micro-py==1.0.3
urllib3==2.2.1
yarl==1.9.4
zstandard==0.22.0
